Allow typed decimal points between operands
Create function decimalPointOkay (script.js, line 149) to check if
a) display.value does not contain a "." or
b) there is an operand (+, -, *, /, (, ) or ^) between the cursor and
the previous decimal point.

If so, the function returns true and doesn't prevent another decimal
point from being typed.

TODO: Call the function successfully for mouse clicking the buttons.