Fix calc zero value bug (issue #76)
I fixed the function 'clear' to set the display value to '0' instead of an empty string.

I also fixed the function 'backspace' to check if the display value is more than 1. If it is, then it removes the last character. It calls ' clear ' if it is 1 or less, as will it if the display value contains letters. Also, I changed the 'letters' variable to be a 'const' instead of a 'var'.

Lastly, I changed the 'backspace' button, which now calls the 'backspace' function.